Undo breaking protobuf changes
Summary: As part of the merge, some protobuf changes from `dev` made it into `master`. Removing those that are considered potential breaking.

Reviewers: fgalilee, mabansal, #peloton, min

Reviewed By: #peloton, min

Subscribers: min, aholler, zhitao, jenkins

Differential Revision: https://code.uberinternal.com/D1267321

+++ jobmgr/job/recovery_test.go
@@ -62,7 +62,7 @@ func TestValidatorWithStore(t *testing.T) {
 	}
 	assert.Nil(t, csStore.CreateJobConfig(context.Background(), jobID, &jobConfig))
 	assert.Nil(t, csStore.CreateJobRuntime(context.Background(), jobID, &job.RuntimeInfo{
-		ConfigVersion: jobConfig.GetRevision().GetVersion(),
+		ConfigVersion: int64(jobConfig.GetRevision().GetVersion()),
 	}, &jobConfig))
 
 	// Create a job with instanceCount 10 but only 5 tasks
